refactor : QueryForObject에서 값이 하나가 아닐 시 에러를 발생하도록 로직 수정
