+++ jdbc/src/main/java/nextstep/jdbc/ResultSetExtractor.java
@@ -5,14 +5,22 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
+import nextstep.jdbc.exception.EmptyResultDataAccessException;
+import nextstep.jdbc.exception.IncorrectResultSizeDataAccessException;
 
 public class ResultSetExtractor {
 
     public static <T> T extract(RowMapper<T> rowMapper, ResultSet rs) throws SQLException {
-        if (rs.next()) {
-            return rowMapper.mapRow(rs);
+        int resultSetColumnCount = getResultSetColumnCount(rs);
+
+        if (resultSetColumnCount == 0) {
+            throw new EmptyResultDataAccessException("데이터가 존재하지 않습니다.");
         }
-        return null;
+        if (resultSetColumnCount > 1) {
+            System.out.println(resultSetColumnCount);
+            throw new IncorrectResultSizeDataAccessException("데이터의 크기가 적절하지 않습니다.");
+        }
+        return rowMapper.mapRow(rs);
     }
 
     public static <T> List<T> extractList(RowMapper<T> rowMapper, ResultSet rs) throws SQLException {
@@ -22,5 +30,12 @@ public class ResultSetExtractor {
             result.add(rowMapper.mapRow(rs));
         }
         return result;
+    }
+
+    private static int getResultSetColumnCount(ResultSet rs) throws SQLException {
+        rs.last();
+        int row = rs.getRow();
+        rs.first();
+        return row;
     }
 }
