+++ jdbc/src/main/java/nextstep/jdbc/exception/EmptyResultDataAccessException.java
@@ -1,0 +1,8 @@
+package nextstep.jdbc.exception;
+
+public class EmptyResultDataAccessException extends RuntimeException{
+
+    public EmptyResultDataAccessException(String message) {
+        super(message);
+    }
+}
